+++ nexus/cmd/hub/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/wandb/wandb/nexus/pkg/hub"
+)
+
+func main() {
+	service := hub.TcpService{}
+	service.Serve()
+}
