+++ nexus/cmd/nexus/main.go
@@ -1,38 +1,34 @@
 package main
 
 import (
-	// "flag"
-	// "fmt"
-	// log "github.com/sirupsen/logrus"
-
-	// "github.com/wandb/wandb/nexus/pkg/server"
-	"github.com/wandb/wandb/nexus/pkg/hub"
+	"flag"
+	log "github.com/sirupsen/logrus"
+	"github.com/wandb/wandb/nexus/pkg/server"
 )
 
 func main() {
-	// portFilename := flag.String("port-filename", "portfile.txt", "filename")
+	/*
+	Main entry point for the nexus service.
+	*/
+	portFilename := flag.String("port-filename", "portfile.txt", "filename")
 
-	// pid := flag.Int("pid", 0, "pid")
-	// debug := flag.Bool("debug", false, "debug")
-	// serveSock := flag.Bool("serve-sock", false, "debug")
-	// serveGrpc := flag.Bool("serve-grpc", false, "debug")
+	pid := flag.Int("pid", 0, "pid")
+	debug := flag.Bool("debug", false, "debug")
+	serveSock := flag.Bool("serve-sock", false, "debug")
+	serveGrpc := flag.Bool("serve-grpc", false, "debug")
 
-	// fmt.Println("Starting 0.0\n")
-	// flag.Parse()
+	flag.Parse()
 
-	// server.InitLogging()
-	//
-	//// log.Println("got", *portFilename, *pid, *debug, *serveSock, *serveGrpc)
-	// log.WithFields(log.Fields{
-	//	"fname":     *portFilename,
-	//	"pid":       *pid,
-	//	"debug":     *debug,
-	//	"serveSock": *serveSock,
-	//	"serveGrpc": *serveGrpc,
-	// }).Debug("Flags")
-	//
-	// server.WandbService(*portFilename)
+	server.InitLogging()
 
-	service := hub.TcpService{}
-	service.Serve()
+	log.WithFields(log.Fields{
+		"fname":     *portFilename,
+		"pid":       *pid,
+		"debug":     *debug,
+		"serveSock": *serveSock,
+		"serveGrpc": *serveGrpc,
+	}).Debug("Flags")
+
+	server.WandbService(*portFilename)
+
 }
